requirements.txt
# AI APIs
openai>=1.0.0
google-genai

# Wav2Vec2 Force Alignment (FREE local alignment)
torch>=2.0.0
torchaudio>=2.0.0

# Web UI
streamlit>=1.30.0

# Google APIs for YouTube upload
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

# Video Processing
moviepy==1.0.3
Pillow>=9.5.0

# Audio / Text-to-Speech
gTTS
pydub

# Web requests
requests
aiohttp

# Environment and CLI
python-dotenv
argparse

# Image processing (for Pexels fallback)
requests

# Optional: Rich console output (nice to have)
rich
